+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Hover information and signature help for magic words in Wikitext
 - Inlay hints for template anonymous parameters in Wikitext
 
-**Fixed**
+**Changed**
 
 - Language services for Wikitext are now provided directly by [WikiParser-Node](https://github.com/bhsd-harry/wikiparser-node), which is consistent with the [VSCode extension](https://github.com/bhsd-harry/vscode-extension-wikiparser)